Externalizing KeccakF1600_Round_vector into its own file
