+++ src/Makefile
@@ -25,11 +25,12 @@ SIMULATOR=spike --isa=rv64gcv_zicntr_zihpm_zvbb --varch=vlen:$(VLEN),elen:64 $(P
 CFLAGS ?= -O2
 
 KECCAK_SRC ?= keccak-vector.c # Keccak-more-compact.c
+KECCAK_WRAPPER_SRC ?= keccak-vector-wrapper.c
 
 keccak.o: $(KECCAK_SRC) 
 	 $(RISCVCC) $(CFLAGS) -c -menable-experimental-extensions -march=rv64gcv$(CLANG_EXTRA_EXTS) $^ -o $@
 
-bench_keccak: Keccak-compact-test.c keccak.o genKAT.c
+bench_keccak: Keccak-compact-test.c $(KECCAK_WRAPPER_SRC) keccak.o genKAT.c
 	 $(RISCVGNUCC) $(CFLAGS) -march=rv64gcv $^ -o $@
 
 sim_bench_keccak: bench_keccak
